+++ st3/mdpopups/__init__.py
@@ -51,7 +51,6 @@ there are helpful errors.</p></div>
 '''
 HL_SETTING = 'mdpopups.use_sublime_highlighter'
 STYLE_SETTING = 'mdpopups.default_style'
-LEGACY_MATCHER_SETTING = 'mdpopups.legacy_color_matcher'
 RE_BAD_ENTITIES = re.compile(r'(&(?!amp;|lt;|gt;|nbsp;)(?:\w+;|#\d+;))')
 
 NODEBUG = 0
@@ -178,8 +177,7 @@ def _get_scheme(scheme):
             if (
                 _is_cache_expired(t) or
                 obj.use_pygments != (not settings.get(HL_SETTING, True)) or
-                obj.default_style != settings.get(STYLE_SETTING, True) or
-                obj.legacy_color_matcher != settings.get(LEGACY_MATCHER_SETTING, False)
+                obj.default_style != settings.get(STYLE_SETTING, True)
             ):
                 obj = None
                 user_css = ''
@@ -546,23 +544,19 @@ def scope2style(view, scope, selected=False, explicit_background=False):
     }
     obj = _get_scheme(view.settings().get('color_scheme'))[0]
     style_obj = obj.guess_style(view, scope, selected, explicit_background)
-    if not obj.legacy_color_matcher:
-        style['color'] = style_obj['foreground']
-        style['background'] = style_obj['background']
-        font = []
-        if style_obj['bold']:
-            font.append('bold')
-        if style_obj['italic']:
-            font.append('italic')
-        if style_obj['underline']:
-            font.append('underline')
-        if style_obj['glow']:
-            font.append('glow')
-        style['style'] = ' '.join(font)
-    else:
-        style['color'] = style_obj.fg_simulated
-        style['background'] = style_obj.bg_simulated
-        style['style'] = style_obj.style
+    style['color'] = style_obj['foreground']
+    style['background'] = style_obj['background']
+    font = []
+    if style_obj['bold']:
+        font.append('bold')
+    if style_obj['italic']:
+        font.append('italic')
+    if style_obj['underline']:
+        font.append('underline')
+    if style_obj['glow']:
+        font.append('glow')
+    style['style'] = ' '.join(font)
+
     return style