+++ docs/src/markdown/about/changelog.md
@@ -10,7 +10,7 @@
   for any API function (except `syntax_highlight`), it will be ignored.
 - **NEW**: Support latest Markdown and Pymdown Extensions. Must use `markdown.extensions.md_in_html` instead of
   `pymdownx.extrarawhtml`.
-- **NEW**: Requires `coloraide` dependency.
+- **NEW**: Better color handling logic.
 - **NEW**: Remove `mdpopups.legacy_color_matcher` option along with legacy color match logic.
 - **FIX**: Fix exception when using Pygments.
 - **FIX**: Fix extra newlines in code blocks.